judge_server/requirement.txt
docker-py>=1.7.2
requests>=2.9.1
six>=1.10.0
websocket-client>=0.35.0
